requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.6.0
torchvision==0.21.0
torchaudio==2.6.0

transformers
accelerate
triton-windows
ffmpeg-python
datasets
soundfile
librosa